Fixed long class code smell adding new one called general magager
