+++ helpers/database_manager.py
@@ -1,15 +1,16 @@
 from pymongo import MongoClient
-import helpers.connection_params as params
-from helpers.file_parser import parse_file, parse_headers
+from .connection_params import CONNECTION_PARAMS as params
+from .file_parser import parse_file, parse_headers
+from .general_manager import GManager
 
-
-class Manager():
+class Manager:
 	def __init__(self):
 		connection = MongoClient(
 	        'mongodb://{user}:{password}@{host}:'
-	        '{port}/{namespace}'.format(**params.CONNECTION_PARAMS)
+	        '{port}/{namespace}'.format(**params)
 	    )
 		self.db = connection.mining_blocks
+		self.gmanager = GManager()
 
 	def insert_new_mineral_deposit(self, data_file):
 		file = open(data_file)
@@ -24,7 +25,7 @@ class Manager():
 	def insert_new_block_model(self, mineral_deposit_name, headers_file):
 		headers = parse_headers(headers_file)
 		mineral_deposit = self.fetch_mineral_deposit(mineral_deposit_name)
-		data_map = self.map_headers(headers, mineral_deposit["minerals"])
+		data_map = self.gmanager.map_headers(headers, mineral_deposit["minerals"])
 		name = ""
 		while name == "":
 			name = input("What would you like to name this block model?\n>>> ")
@@ -53,13 +54,13 @@ class Manager():
 			print("\t{}. {}".format(index + 1, grade_units[index]))
 		for grade in grades_data_map:
 			print("{}: ".format(grade), end = '')
-			unit_index = self.get_valid_index(grade_units) - 1
+			unit_index = self.gmanager.get_valid_index(grade_units) - 1
 			my_units[grades_data_map[grade]] = unit_index
 		for item in data:
 			document = { "mineral_deposit": mineral_deposit, "block_model": block_model }
 			for index in range(amount_headers):
 				if headers[index] in grades_data_map.values():
-					transformed = self.grade_to_percentage(item[index], item[weight_column_index], grade_units[my_units[headers[index]]])
+					transformed = self.gmanager.grade_to_percentage(item[index], item[weight_column_index], grade_units[my_units[headers[index]]])
 					item[index] = transformed
 				document[headers[index]] = item[index]
 			data_array.append(document)
@@ -76,46 +77,6 @@ class Manager():
 	def remove_all_blocks_from_block_model(self, mineral_deposit, block_model):
 		return self.db.blocks.remove({ "mineral_deposit_name": mineral_deposit, "block_model": block_model })
 
-	def map_headers(self, headers, minerals):
-		data_map = {}
-		print("You have entered the following headers for your block model:")
-		for index in range(len(headers)):
-			print("\t{}. {}".format(index + 1, headers[index]))
-		print("Please tell me, which one of those hearders refers to:")
-		relevant_headers = ['x', 'y', 'z', 'weight']
-		for header in relevant_headers:
-			print("{}: ".format(header), end = '')
-			response = self.get_valid_index(headers) - 1
-			data_map[header] = headers[response]
-		data_map["grade"] = {}
-		for mineral in minerals:
-			print("{} grade: ".format(mineral), end = '')
-			response = self.get_valid_index(headers) - 1
-			data_map["grade"][mineral] = headers[response]
-		return data_map
-
-	def get_valid_index(self, data):
-		response = -1
-		while response > len(data) or response <= 0:
-			try:
-				response = int(input())
-			except:
-				f = input("exit? y/n\n")
-				if f == 'y':
-					exit()
-				continue
-		return response
-
-
-	def grade_to_percentage(self, value, weight, unit):
-		if unit == 'tonn':
-			return value / weight
-		elif unit == 'oz/tonn':
-			return value * 32000 / weight
-		elif unit == 'ppm':
-			return value / 10000
-		else:
-			return value
 
 	def get_max_x_y_z_value(self, mineral_deposit_name, block_model_name):
 		block_model = self.db.block_models.find_one({"mineral_deposit_name": mineral_deposit_name, "name": block_model_name})
